requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0
ultralytics>=8.0.0

# Machine Learning & Data Science
scikit-learn>=1.3.0
numpy>=1.24.0

# Model Components
timm>=0.9.0

# Visualization & Progress
matplotlib>=3.7.0
tqdm>=4.65.0

# Web Interface
streamlit>=1.28.0

# Fix protobuf compatibility issues
protobuf==3.20.3

# Additional dependencies that might be needed
lxml>=4.9.0
PyYAML>=6.0 